fuch

+++ srcs/ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: midounhocine <mdho@hotmail.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/12 13:24:25 by midounhoc         #+#    #+#             */
-/*   Updated: 2018/08/13 14:07:07 by midounhoc        ###   ########.fr       */
+/*   Updated: 2019/02/11 14:07:04 by midounhoc        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void    *ft_memccpy(void *dst, const void *src, int c,  size_t n)
 {
     unsigned char   *cp_dst;
     unsigned char   *cp_src;
-    int             i;
+    size_t          i;
 
     cp_dst = (unsigned char *)dst;
     cp_src = (unsigned char *)src;
