+++ srcs/ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: hmidoun <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/10 01:16:26 by hmidoun           #+#    #+#             */
-/*   Updated: 2018/08/13 21:14:50 by midounhoc        ###   ########.fr       */
+/*   Updated: 2019/02/11 14:10:18 by midounhoc        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strncat(char *dest, const char *src, size_t nb)
 {
 	int i;
-	int j;
+	size_t j;
 
 	j = 0;
 	i = ft_strlen(dest);
