+++ srcs/ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: midounhocine <mdho@hotmail.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/22 15:14:29 by midounhoc         #+#    #+#             */
-/*   Updated: 2019/01/22 15:22:21 by midounhoc        ###   ########.fr       */
+/*   Updated: 2019/02/11 14:12:24 by midounhoc        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char* ft_strnew(size_t size)
 {
-    char* str;
-    int i = 0;
+    char*   str;
+    size_t  i = 0;
     str = (char*)malloc(size * sizeof(char));
     if(str == NULL) return NULL;
     while(i < size)
