+++ srcs/ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: midounhocine <mdho@hotmail.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/22 16:38:20 by midounhoc         #+#    #+#             */
-/*   Updated: 2019/01/22 16:42:47 by midounhoc        ###   ########.fr       */
+/*   Updated: 2019/02/11 14:11:37 by midounhoc        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 char* ft_strsub(char const *s, unsigned int start, size_t len)
 {
     char* str;
-    int i = 0;
+    size_t i = 0;
     str = (char*)malloc(sizeof(char) * len);
     if(!str) return(NULL);
     while(i < len)
